// src/app/_dto/TblHAuthMaster.dto.ts
import { TbldAuthMasterCo } from "./TbldAuthMasterCo.dto";

export class TblHAuthMaster {
    HAuthMaste_SysID: number;
    HAuthMaste_Code: string;
    HAuthMaste_Name: string;
    HAuthMaste_AcDe_Yn: string;
    HAuthMaste_AcDe_SysID: number;
    HAuthMaste_AcDe_Code: string;
    HAuthMaste_AcDe_Name: string;
    
    TbldAuthMasterCo :TbldAuthMasterCo[] = []

}